Reference

Contents

Reference

Contents

Index

Index

LorentzVectorBase.PtEtaPhiM

LorentzVectorBase.XYZE

LorentzVectorBase.boost_beta

LorentzVectorBase.boost_gamma

LorentzVectorBase.coordinate_names

LorentzVectorBase.coordinate_system

LorentzVectorBase.cos_phi

LorentzVectorBase.cos_theta

LorentzVectorBase.energy

LorentzVectorBase.eta

LorentzVectorBase.mass

LorentzVectorBase.mass2

LorentzVectorBase.minus_component

LorentzVectorBase.mt

LorentzVectorBase.mt2

LorentzVectorBase.phi

LorentzVectorBase.plus_component

LorentzVectorBase.polar_angle

LorentzVectorBase.pt

LorentzVectorBase.pt2

LorentzVectorBase.px

LorentzVectorBase.py

LorentzVectorBase.pz

LorentzVectorBase.rapidity

LorentzVectorBase.sin_phi

LorentzVectorBase.spatial_magnitude

LorentzVectorBase.spatial_magnitude2

LorentzVectorBase.PtEtaPhiM — Type

PtEtaPhiM <: AbstractCoordinateSystem

Cylindrical coordinate system for four-momenta. Using this requires the implementation of the following interface functions:

pt(::CustomFourMomentum)
eta(::CustomFourMomentum)
phi(::CustomFourMomentum)
mass(::CustomFourMomentum)

LorentzVectorBase.XYZE — Type

XYZE <: AbstractCoordinateSystem

Cartesian coordinate system for four-momenta. Using this requires the implementation of the following interface functions:

px(::CustomFourMomentum)
py(::CustomFourMomentum)
pz(::CustomFourMomentum)
energy(::CustomFourMomentum)

LorentzVectorBase.boost_beta — Method

boost_beta(::XYZE, mom )

Return spatialmagnitude of the beta vector for a given four-momentum, i.e. the spatialmagnitude of the four-momentum divided by its 0-component.

Example

If (px,py,pz,E) is a four-momentum, this is equivalent to sqrt(px^2 + py^2 + pz^2)/E.

LorentzVectorBase.boost_gamma — Method

boost_gamma(::XYZE,mom)

Return the relativistic gamma factor for a given four-momentum, i.e. the inverse square root of one minus the beta vector squared.

Example

If (px,py,pz,E) is a four-momentum with beta vector β, this is equivalent to 1/sqrt(1- β^2).

LorentzVectorBase.coordinate_names — Function

coordinate_names(::AbstractCoordinateSystem)::Tuple{Symbols}

LorentzVectorBase.coordinate_system — Function

coordinate_system(::MomType)::CS where {CS <: AbstractCoordinateSystem}

LorentzVectorBase.cos_phi — Method

cos_phi(::XYZE, mom)

Return the cosine of the azimuthal angle for a given four-momentum.

Note

This is an equivalent but faster version of cos(azimuthal_angle(mom)); see phi.

LorentzVectorBase.cos_theta — Method

cos_theta(::XYZE, mom)

Return the cosine of the theta angle for a given four-momentum.

Note

This is an equivalent but faster version of cos(polar_angle(::XYZE, mom)); see polar_angle.

LorentzVectorBase.energy — Function

energy(mom)

TBW

LorentzVectorBase.eta — Method

eta(::XYZE, mom)

TBW

LorentzVectorBase.mass — Method

mass(::XYZE,mom)

Return the the invariant mass of a given four-momentum, i.e. the square root of the minkowski dot with itself.

Example

If (px,py,pz,E) is a four-momentum, this is equivalent to sqrt(E^2 - (px^2 + py^2 + pz^2)).

Note

If the squared invariant mass m2 is negative, -sqrt(-m2) is returned.

LorentzVectorBase.mass2 — Method

mass2(::XYZE,mom)

Return the squared invariant mass of a given four-momentum, i.e. the minkowski dot with itself.

Example

If (px,py,pz,E) is a four-momentum, this is equivalent to E^2 - (px^2 + py^2 + pz^2).

LorentzVectorBase.minus_component — Method

minus_component(::XYZE, mom)

Return the minus component for a given four-momentum in light-cone coordinates.

Example

If (px,py,pz,E) is a four-momentum, this is equivalent to (E-pz)/2.

Note

The light-cone coordinates are defined w.r.t. to the 3-axis.

Warning

The definition `p^- := (E - p_z)/2 differs from the usual definition of light-cone coordinates in general relativity.

LorentzVectorBase.mt — Method

mt(::XYZE,mom)

Return the transverse momentum for a given four-momentum, i.e. the square root of its squared transverse mass.

Example

If (px,py,pz,E) is a four-momentum, this is equivalent to sqrt(E^2 - pz^2).

Note

The transverse components are defined w.r.t. to the 3-axis.

Note

If the squared transverse mass mt2 is negative, -sqrt(-mt2) is returned.

LorentzVectorBase.mt2 — Method

mt2(::XYZE, mom)

Return the squared transverse mass for a given four-momentum, i.e. the difference of its squared 0- and 3-component.

Example

If (px,py,pz,E) is a four-momentum, this is equivalent to E^2 - pz^2.

Note

The transverse components are defined w.r.t. to the 3-axis.

LorentzVectorBase.phi — Method

phi(::XYZE, mom)

Return the phi angle for a given four-momentum, i.e. the azimuthal angle of its spatial components in spherical coordinates.

Example

If (px,py,pz,E) is a four-momentum, this is equivalent to atan(py,px).

Note

The spherical coordinates are defined w.r.t. to the 3-axis.

LorentzVectorBase.plus_component — Method

plus_component(::XYZE, mom)

Return the plus component for a given four-momentum in light-cone coordinates.

Example

If (px,py,pz,E) is a four-momentum, this is equivalent to (E+pz)/2.

Note

The light-cone coordinates are defined w.r.t. to the 3-axis.

Warning

The definition `p^+ := (E + p_z)/2 differs from the usual definition of light-cone coordinates in general relativity.

LorentzVectorBase.polar_angle — Method

polar_angle(::XYZE,mom)

Return the theta angle for a given four-momentum, i.e. the polar angle of its spatial components in spherical coordinates.

Example

If (px,py,pz,E) is a four-momentum with spatial_magnitude rho, this is equivalent to arccos(pz/rho), which is also equivalent to arctan(sqrt(px^2+py^2)/pz).

Note

The spherical coordinates are defined w.r.t. to the 3-axis.

LorentzVectorBase.pt — Method

pt(::XYZE, mom)

Return the transverse momentum for a given four-momentum, i.e. the spatial_magnitude of its transverse components.

Example

If (px,py,pz,E) is a four-momentum, this is equivalent to sqrt(px^2 + py^2).

Note

The transverse components are defined w.r.t. to the 3-axis.

LorentzVectorBase.pt2 — Method

pt2(::XYZE,mom)

Return the squared transverse momentum for a given four-momentum, i.e. the sum of its squared 1- and 2-component.

Example

If (px,py,pz,E) is a four-momentum, this is equivalent to px^2 + py^2.

Note

The transverse components are defined w.r.t. to the 3-axis.

LorentzVectorBase.px — Function

px(mom)

TBW

LorentzVectorBase.py — Function

py(mom)

TBW

LorentzVectorBase.pz — Function

pz(mom)

TBW

LorentzVectorBase.rapidity — Method

rapidity(::XYZE, mom)

Return the rapidity for a given four-momentum.

Example

If (px,py,pz,E) is a four-momentum, this is equivalent to 0.5*log((E+pz)/(E-pz)).

Note

The transverse components are defined w.r.t. to the 3-axis.

LorentzVectorBase.sin_phi — Method

sin_phi(::XYZE,mom)

Return the sine of the azimuthal angle for a given four-momentum.

Note

This is an equivalent but faster version of sin(azimuthal_angle(mom)); see phi.

LorentzVectorBase.spatial_magnitude — Method

spatial_magnitude(::XYZE,mom)

Return the spatial_magnitude of a given four-momentum, i.e. the euklidian norm spatial components.

Example

If (px,py,pz,E) is a four-momentum, this is equivalent to sqrt(px^2 + py^2 + pz^2).

Warning

This function differs from a similar function for the TLorentzVector used in the famous ROOT library.

LorentzVectorBase.spatial_magnitude2 — Method

spatial_magnitude2(::XYZE, mom)

Return the square of the spatial_magnitude of a given four-momentum, i.e. the sum of the squared spatial components.

Example

If (px,py,pz,E) is a four-momentum, this is equivalent to px^2+ py^2 + pz^2.

Warning

This function differs from a similar function for the TLorentzVector used in the famous ROOT library.